requirements.txt
beautifulsoup4==4.13.4
pandas==2.3.1
plotly==5.24.1
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.42.2
matplotlib
seaborn